Split out WePay::Client
- paves way for more OO refactoring
